requirements-video-ai.txt
# Core AI and LangChain (optional, fallback to Gemini if not available)
langchain==0.0.350
openai==1.3.7

# Google AI (primary)
google-generativeai==0.3.2

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google APIs (optional, for enhanced metadata)
google-api-python-client==2.111.0

# YouTube processing (essential)
youtube-transcript-api==0.6.1

# Vector stores (optional, for advanced search)
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Text processing
nltk==3.8.1
